requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
pillow>=10.0.0
pydantic>=2.6.0
fastapi>=0.110.0
uvicorn>=0.29.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
pyyaml>=6.0.0
requests>=2.31.0
mlflow>=2.11.0
opencv-python>=4.9.0.80
